등급 처리 기능 추가 예정! 누락된 파일을 추가했습니다

+++ backend/src/main/java/com/cw/cwu/domain/Grade.java
@@ -17,7 +17,7 @@ public class Grade {
     private Long gradeId;
 
     @Column(name = "enrollment_id")
-    private Long enrollmentId;
+    private int enrollmentId;
 
     @Column(name = "grade_grade")
     private StudentGrade grade;
@@ -43,7 +43,4 @@ public class Grade {
             default -> StudentGrade.valueOf(db);
         };
     }
-
-
-
 }
